+++ Goo.Svg/Svg.cs
@@ -14,30 +14,18 @@ public static class Svg
     public static VectorAsset Parse(string source)
     {
         ArgumentNullException.ThrowIfNull(source);
-        return LoadCompiled(() => global::Goo.SvgCompiler.SvgCompiler.CompileText(source));
+        return SvgParser.ParseText(source);
     }
 
     public static VectorAsset Load(Stream stream)
     {
         ArgumentNullException.ThrowIfNull(stream);
-        return LoadCompiled(() => global::Goo.SvgCompiler.SvgCompiler.CompileStream(stream, "<stream>"));
+        return SvgParser.LoadStream(stream, "<stream>");
     }
 
     public static VectorAsset Load(string path)
     {
         ArgumentException.ThrowIfNullOrWhiteSpace(path);
-        return LoadCompiled(() => global::Goo.SvgCompiler.SvgCompiler.CompileFile(path));
-    }
-
-    private static VectorAsset LoadCompiled(Func<byte[]> compile)
-    {
-        try
-        {
-            return VectorAsset.Load(compile());
-        }
-        catch (global::Goo.SvgCompiler.SvgCompileException exception)
-        {
-            throw new SvgParseException(exception.Message, exception);
-        }
+        return SvgParser.LoadFile(path);
     }
 }